+++ tests/test_program_generation.py
@@ -1,14 +1,13 @@
 from __future__ import annotations
 
-import json
+import importlib.util
 from pathlib import Path
 
-import orjson
 import pytest
-import serde.json
+from rdflib import Graph, compare
 
 from drepr.models.prelude import DRepr, OutputFormat
-from drepr.planning.class_map_plan import BlankSubject, ClassesMapExecutionPlan, Subject
+from drepr.planning.class_map_plan import ClassesMapExecutionPlan
 from drepr.program_generation.main import FileOutput, MemoryOutput, gen_program
 from tests.conftest import DatasetExample
 
@@ -20,7 +19,7 @@ from tests.conftest import DatasetExample
         "mineral_site/missing_values",
         "mineral_system/misspath_autoalign",
         "resource_categories",
-    ][-1:],
+    ],
 )
 def test_program_generation(
     name, example_datasets: dict[str, DatasetExample], tmp_path: Path
@@ -55,3 +54,25 @@ def test_program_generation(
         pytest.fail(
             "The generated program is different from the expected program. Please check the file: program/write_to_file.auto.py"
         )
+
+    # now invoke the generated program and compare the output
+    spec = importlib.util.spec_from_file_location(
+        "drepr_prog", (ds.cwd / f"program/write_to_str.py")
+    )
+    assert spec is not None and spec.loader is not None
+    module = importlib.util.module_from_spec(spec)
+    spec.loader.exec_module(module)
+
+    output = module.main(*[ds.resources[r.id] for r in model.resources])
+    # (ds.cwd / "tmp.ttl").write_text(output)
+    gold_g = Graph()
+    gold_g.parse(data=output)
+
+    pred_g = Graph()
+    pred_g.parse(data=ds.output.read_text())
+    # x, y, z = compare.graph_diff(gold_g, pred_g)
+    # print(list(x), list(y), list(z))
+
+    assert compare.isomorphic(
+        gold_g, pred_g
+    ), "The output of the generated program is different from the expected output"